rename view provider
